+++ src/main/java/com/nhnacademy/eggplantdeliveryschedule/writer/DeliveryLocationChangeWriter.java
@@ -1,8 +1,7 @@
-package com.nhnacademy.eggplantdeliveryschedule.entity.status.writer;
+package com.nhnacademy.eggplantdeliveryschedule.writer;
 
 import com.nhnacademy.eggplantdeliveryschedule.entity.DeliveryInfo;
 import com.nhnacademy.eggplantdeliveryschedule.entity.Location;
-import com.nhnacademy.eggplantdeliveryschedule.entity.status.Status;
 import com.nhnacademy.eggplantdeliveryschedule.exception.NotFoundDeliveryInfoException;
 import com.nhnacademy.eggplantdeliveryschedule.repository.DeliveryInfoRepository;
 import com.nhnacademy.eggplantdeliveryschedule.repository.LocationRepository;
@@ -10,31 +9,26 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import org.springframework.batch.core.StepExecution;
-import org.springframework.batch.core.annotation.BeforeStep;
 import org.springframework.batch.core.configuration.annotation.StepScope;
-import org.springframework.batch.item.ExecutionContext;
 import org.springframework.batch.item.ItemWriter;
 import org.springframework.stereotype.Component;
 
 /**
- * @author : 조재철
- * @since 1.0
+ * DeliveryLocationChangeReader 에서 읽은 데이터의 정보를 처리하기 위한 Writer 클래스 입니다.
+ *
+ * @author 조재철, 김훈민
+ * @version 1.0.0
  */
 @Component
 @StepScope
 @RequiredArgsConstructor
-public class ReadyToDeliveringWriter implements ItemWriter<List<String>> {
+public class DeliveryLocationChangeWriter implements ItemWriter<List<String>> {
 
-    private StepExecution stepExecution;
     private final DeliveryInfoRepository deliveryInfoRepository;
     private final LocationRepository locationRepository;
 
     @Override
     public void write(List<? extends List<String>> trackingNoChunkList) {
-        ExecutionContext executionContext = this.stepExecution.getExecutionContext();
-        executionContext.put("trackingNoChunkList", trackingNoChunkList);
-
         List<String> trackingNoList = trackingNoChunkList.stream()
                                                          .flatMap(List::stream)
                                                          .collect(Collectors.toList());
@@ -42,21 +36,14 @@ public class ReadyToDeliveringWriter implements ItemWriter<List<String>> {
         for (String trackingNo : trackingNoList) {
             DeliveryInfo deliveryInfo = deliveryInfoRepository.findById(trackingNo)
                                                               .orElseThrow(NotFoundDeliveryInfoException::new);
-            deliveryInfo.updateStatus(Status.DELIVERING);
-
             Location location = Location.builder()
-                                        .pk(new Location.Pk(1L, deliveryInfo.getTrackingNo()))
+                                        .pk(new Location.Pk(2L, deliveryInfo.getTrackingNo()))
                                         .deliveryInfo(deliveryInfo)
                                         .arrivalTime(LocalDateTime.now())
                                         .build();
 
             locationRepository.save(location);
         }
-    }
-
-    @BeforeStep
-    public void saveStepExecution(StepExecution stepExecution) {
-        this.stepExecution = stepExecution;
     }
 
 }
