+++ src/test/java/com/nhnacademy/eggplantdeliveryschedule/repository/CustomDeliveryInfoRepositoryTest.java
@@ -2,7 +2,7 @@ package com.nhnacademy.eggplantdeliveryschedule.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.nhnacademy.eggplantdeliveryschedule.entity.status.writer.DeliveryLocationChangeWriter;
+import com.nhnacademy.eggplantdeliveryschedule.writer.DeliveryLocationChangeWriter;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
